+++ lib/cubes.dart
@@ -5,5 +5,6 @@ export 'package:cubes/src/cube_builder.dart';
 export 'package:cubes/src/injector.dart';
 export 'package:cubes/src/localization/cubes_localization_delegate.dart';
 export 'package:cubes/src/localization/strings_location.dart';
+export 'package:cubes/src/observable_list.dart';
 export 'package:cubes/src/observable_value.dart';
 export 'package:cubes/src/util/extensions/ext.dart';